+++ dcb/src/main/java/org/olf/dcb/core/interaction/alma/AlmaHostLmsClient.java
@@ -17,6 +17,8 @@ import org.olf.dcb.core.interaction.folio.MaterialTypeToItemTypeMappingService;
 import org.olf.dcb.core.svc.LocationToAgencyMappingService;
 import services.k_int.interaction.alma.AlmaApiClient;
 import services.k_int.interaction.alma.types.*;
+import services.k_int.interaction.alma.types.error.AlmaError;
+import services.k_int.interaction.alma.types.error.AlmaException;
 import services.k_int.interaction.alma.types.holdings.AlmaHolding;
 import services.k_int.interaction.alma.types.items.*;
 import services.k_int.interaction.alma.types.userRequest.*;
@@ -73,6 +75,8 @@ public class AlmaHostLmsClient implements HostLmsClient {
 		= stringPropertyDefinition("shelf-location", "Shelf location for this ALMA system", FALSE);
 	private static final HostLmsPropertyDefinition PICKUP_CIRC_DESK_SETTING
 		= stringPropertyDefinition("pickup-circ-desk", "Pickup circ desk for this ALMA system", FALSE);
+	private static final HostLmsPropertyDefinition USER_IDENTIFIER
+		= stringPropertyDefinition("user-identifier", "User identifier to find patron", FALSE);
 
 	private static final HostLmsPropertyDefinition PICKUP_LIBRARY_SETTING
 		= stringPropertyDefinition("pickup-location-library", "Default pickup library for this ALMA system", FALSE);
@@ -286,43 +290,59 @@ public class AlmaHostLmsClient implements HostLmsClient {
 
 	@Override
 	public Mono<Patron> findVirtualPatron(org.olf.dcb.core.model.Patron patron) {
+		log.info("Finding virtual patron {}", patron);
 
-		// find user by external_id (DCB patron uniqueId)
-		// this is assuming when a virtual patron is created the external_id is set with DCBs patron uniqueId
 		final var uniqueId = getValueOrNull(patron, org.olf.dcb.core.model.Patron::determineUniqueId);
-		final var uniqueId2 = getValueOrNull(patron, org.olf.dcb.core.model.Patron::getHomeLibraryCode);
 
-		log.info(uniqueId, uniqueId2);
-		return client.getUsersByExternalId(uniqueId)
-			.map(almaUserList -> {
+		if (uniqueId == null) {
+			return Mono.error(new IllegalArgumentException("Unable to find uniqueId for virtual patron"));
+		}
 
-				final var users = getValueOrNull(almaUserList, AlmaUserList::getUser);
-				final var usersSize = (users != null) ? users.size() : 0;
-
-				if (usersSize < 1) {
-					log.warn("No virtual Patron found.");
-
-					throw VirtualPatronNotFound.builder()
-						.withDetail(usersSize + " records found")
-						.with("uniqueId", uniqueId)
-						.with("Response", almaUserList)
-						.build();
+		// this relies on implementation relies on alma finding the user by our uniqueId
+		// if we have created a user with a user identifier correctly we should be good
+		// the user will need to have the user identifier enabled
+		return client.getAlmaUserByUserId(uniqueId)
+			.doOnNext(almaUser -> log.info("Found virtual patron with uniqueId: {}", uniqueId))
+			.map(this::almaUserToPatron)
+			.onErrorResume(e -> {
+				if (isVirtualPatronNotFoundError(e)) {
+					throw createVirtualPatronNotFoundException(uniqueId, e);
 				}
-
-				if (usersSize > 1) {
-					log.error("More than one virtual patron found: {}", almaUserList);
-
-					throw MultipleVirtualPatronsFound.builder()
-						.withDetail(usersSize + " records found")
-						.with("uniqueId", uniqueId)
-						.with("Response", almaUserList)
-						.build();
-				}
-
-				final var onlyUser = users.get(0);
-
-				return almaUserToPatron(onlyUser);
+				return Mono.error(e);
 			});
+	}
+
+	private static final int BAD_REQUEST_STATUS = 400;
+	private static final String VIRTUAL_PATRON_NOT_FOUND_ERROR_CODE = "401861";
+
+	private boolean isVirtualPatronNotFoundError(Throwable e) {
+		if (!(e instanceof AlmaException almaException)) {
+			return false;
+		}
+
+		if (almaException.getStatus().getCode() != BAD_REQUEST_STATUS) {
+			return false;
+		}
+
+		try {
+			List<AlmaError> errors = almaException.getErrorResponse()
+				.getErrorList()
+				.getError();
+
+			return errors.stream()
+				.anyMatch(error -> VIRTUAL_PATRON_NOT_FOUND_ERROR_CODE.equals(error.getErrorCode()));
+		} catch (Exception ex) {
+			log.error("Unable to determine if virtual patron not found error", ex);
+			return false;
+		}
+	}
+
+	private VirtualPatronNotFound createVirtualPatronNotFoundException(String uniqueId, Throwable cause) {
+		return VirtualPatronNotFound.builder()
+			.withDetail("No records found")
+			.with("uniqueId", uniqueId)
+			.with("Response", cause.toString())
+			.build();
 	}
 
 	// Static create patron defaults
@@ -332,21 +352,18 @@ public class AlmaHostLmsClient implements HostLmsClient {
 	private static final String STATUS_ACTIVE = "ACTIVE";
 	private static final String ACCOUNT_TYPE_EXTERNAL = "EXTERNAL";
 	private static final String ID_TYPE_BARCODE = "BARCODE";
+	private static final String ID_TYPE_INST_ID = "INST_ID";
 
 	@Override
 	public Mono<String> createPatron(Patron patron) {
 
-		final var homeIdentityLocalId = extractExternalId(patron);
 		final var firstName = extractFirstName(patron);
 		final var lastName = extractLastName(patron);
 		final var externalId = extractExternalId(patron);
-		final var primaryId = extractPrimaryId(patron);
-		log.info("Alma patron creation starts with this patron {} and this homeIdentityLocalId {}, this externalId {}, and this primary ID {}", patron, homeIdentityLocalId, externalId, primaryId);
 
-		// We need the primary ID  here to build our Alma user with it, but I don't think we have it
-
-		List<UserIdentifier> userIdentifiers = createUserIdentifiers(patron);
-		AlmaUser almaUser = buildAlmaUser(primaryId, firstName, lastName, externalId, userIdentifiers, primaryId);
+		UserIdentifiers userIdentifiers = createUserIdentifiers(patron);
+		AlmaUser almaUser = buildAlmaUser(firstName, lastName, externalId, userIdentifiers);
+		log.info("Attempting to create a patron for Alma with Patron: {}, alma user: {} and user identifiers {}. First name is {}, last name is {}", patron, almaUser, userIdentifiers, firstName,lastName);
 
 		return determinePatronType(patron)
 			.flatMap(patronType -> {
@@ -355,18 +372,9 @@ public class AlmaHostLmsClient implements HostLmsClient {
 				return Mono.from(client.createPatron(almaUser))
 					.flatMap(returnedUser -> {
 						log.info("Created alma user {}", returnedUser);
-						return Mono.just(homeIdentityLocalId);
+						return Mono.just(returnedUser.getPrimary_id());
 					});
 			});
-	}
-
-	private String extractPatronId(Patron patron) {
-
-		if (patron.getLocalId() != null && !patron.getLocalId().isEmpty()) {
-			return patron.getLocalId().get(0);
-		}
-
-		return null;
 	}
 
 	private String extractFirstName(Patron patron) {
@@ -395,47 +403,50 @@ public class AlmaHostLmsClient implements HostLmsClient {
 	}
 
 
-	private String extractPrimaryId(Patron patron) {
-		if (patron.getUniqueIds() != null && !patron.getUniqueIds().isEmpty()) {
-			return patron.getUniqueIds().get(0);
-		}
-		return null;
-	}
+	private UserIdentifiers createUserIdentifiers(Patron patron) {
+		final var identifierType = USER_IDENTIFIER.getOptionalValueFrom(hostLms.getClientConfig(), ID_TYPE_INST_ID);
+		final String externalId = extractExternalId(patron);
 
-	private List<UserIdentifier> createUserIdentifiers(Patron patron) {
-		if (patron.getLocalBarcodes() != null) {
-			var userIdentifiers = patron.getLocalBarcodes().stream()
-				.map(value -> UserIdentifier.builder()
+		// Validate that we have at least one identifier source
+		if ((patron.getLocalBarcodes() == null || patron.getLocalBarcodes().isEmpty()) && externalId == null) {
+			throw new IllegalArgumentException("Cannot create user identifiers: patron has no barcodes or external ID");
+		}
+
+		List<UserIdentifier> identifiers = new ArrayList<>();
+
+		// Add barcode identifiers
+		if (patron.getLocalBarcodes() != null && !patron.getLocalBarcodes().isEmpty()) {
+			patron.getLocalBarcodes().stream()
+				.filter(Objects::nonNull) // Guard against null barcodes in the list
+				.map(barcode -> UserIdentifier.builder()
 					.id_type(WithAttr.builder().value(ID_TYPE_BARCODE).build())
-					.value(value)
+					.value(barcode)
 					.build())
-				.collect(Collectors.toList());
-
-			userIdentifiers.add(UserIdentifier.builder()
-				.id_type(WithAttr.builder().value("dcb_unique_id").build())
-				.value(extractExternalId(patron)).build());
-
-			userIdentifiers.add(UserIdentifier.builder()
-				.id_type(WithAttr.builder().value("home_identity_local_id").build())
-				.value(extractPatronId(patron)).build());
-
-			return userIdentifiers;
+				.forEach(identifiers::add);
 		}
-		return null;
+
+		// Add external ID identifier (if available and not already added as barcode)
+		if (externalId != null) {
+			identifiers.add(UserIdentifier.builder()
+				.id_type(WithAttr.builder().value(identifierType).build())
+				.value(externalId)
+				.build());
+		}
+		log.info("Identifiers {}", identifiers);
+		return UserIdentifiers.builder().identifiers(identifiers).build();
 	}
 
-	private AlmaUser buildAlmaUser(String patronId, String firstName, String lastName,
-																 String externalId, List<UserIdentifier> userIdentifiers, String primaryId) {
-		log.info("Building an alma user with patronId {}, externalId {}, identifiers {}, primaryId {}", patronId, externalId, userIdentifiers, primaryId);
+	private AlmaUser buildAlmaUser(String firstName, String lastName, String externalId,
+																 UserIdentifiers userIdentifiers) {
 		return AlmaUser.builder()
 			.record_type(CodeValuePair.builder().value(RECORD_TYPE_PUBLIC).build())
 			.first_name(firstName)
 			.last_name(lastName)
 			.status(CodeValuePair.builder().value(STATUS_ACTIVE).build())
 			.is_researcher(Boolean.FALSE)
-			.user_identifiers(userIdentifiers)
-			.external_id(externalId) // DCB Patron ID for home library system
-			.primary_id(primaryId)
+			.identifiers(userIdentifiers)
+			.external_id(externalId)
+			.primary_id(externalId) // Workaround: to be removed once we understand where the primary_id is being lost
 			.account_type(CodeValuePair.builder().value(ACCOUNT_TYPE_EXTERNAL).build())
 			.build();
 	}
@@ -491,7 +502,7 @@ public class AlmaHostLmsClient implements HostLmsClient {
 					.last_name(returnedUser.getLast_name())
 					.status(returnedUser.getStatus())
 					.is_researcher(returnedUser.getIs_researcher())
-					.user_identifiers(returnedUser.getUser_identifiers())
+					.identifiers(returnedUser.getIdentifiers())
 					.external_id(returnedUser.getExternal_id())
 					.account_type(returnedUser.getAccount_type())