+++ dcb/src/main/java/services/k_int/interaction/alma/types/UserIdentifier.java
@@ -1,5 +1,6 @@
 package services.k_int.interaction.alma.types;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import io.micronaut.serde.annotation.Serdeable;
 
 import lombok.AllArgsConstructor;
@@ -11,10 +12,10 @@ import lombok.ToString;
 @Data
 @AllArgsConstructor
 @Builder
-@ToString(onlyExplicitlyIncluded = true)
 @Serdeable
 public class UserIdentifier {
 	// BARCODE, INST_ID, ILLIAD_ID, GoogleScholarID, ORCID,...
+	@JsonProperty("id_type")
 	WithAttr id_type;
 	String value;
 	String note;