+++ dcb/src/main/java/services/k_int/interaction/alma/types/UserIdentifiers.java
@@ -1,17 +1,21 @@
 package services.k_int.interaction.alma.types;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
 import io.micronaut.serde.annotation.Serdeable;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 import lombok.ToString;
+
+import java.util.List;
 
 @Data
 @AllArgsConstructor
 @Builder
+@ToString(onlyExplicitlyIncluded = true)
 @Serdeable
-public class WithAttr {
-        String value;
+public class UserIdentifiers {
+	@ToString.Include
+	@JsonProperty("user_identifier")
+	private List<UserIdentifier> identifiers;
 }